pyproject.toml
[project]
name = "daily-bets"
version = "0.1.0"
description = "Gets daily bets"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.391",
    "ruff>=0.8.4",
]
